+++ src/EntityGraphQL/Schema/MutationField.cs
@@ -37,12 +37,27 @@ namespace EntityGraphQL.Schema
                     AddInputTypesInArguments(schema, autoAddInputTypes, item.PropertyType);
 
                 }
-                foreach (var item in ArgumentsType.GetFields())
+                foreach (var item in ArgumentsType.GetFields(BindingFlags.Instance | BindingFlags.Public))
                 {
                     if (GraphQLIgnoreAttribute.ShouldIgnoreMemberFromInput(item))
                         continue;
                     Arguments.Add(fieldNamer(item.Name), ArgType.FromField(schema, item, null, fieldNamer));
                     AddInputTypesInArguments(schema, autoAddInputTypes, item.FieldType);
+                }
+            }
+            else
+            {
+                foreach (var item in method.GetParameters())
+                {
+                    if (GraphQLIgnoreAttribute.ShouldIgnoreMemberFromInput(item))
+                        continue;
+
+                    var inputType = item.ParameterType.GetEnumerableOrArrayType() ?? item.ParameterType;
+                    if (item.ParameterType.IsPrimitive || schema.HasType(inputType))
+                    {
+                        Arguments.Add(fieldNamer(item.Name), ArgType.FromParameter(schema, item, null, fieldNamer));
+                        AddInputTypesInArguments(schema, autoAddInputTypes, item.ParameterType);
+                    }
                 }
             }
         }
@@ -51,7 +66,7 @@ namespace EntityGraphQL.Schema
         {
             var inputType = propType.GetEnumerableOrArrayType() ?? propType;
             if (autoAddInputTypes && !schema.HasType(inputType))
-                schema.AddInputType(inputType, inputType.Name, null).AddAllFields(true);
+                schema.AddInputType(inputType, inputType.Name, null).AddAllFields();
         }
 
         public void Deprecate(string reason)
@@ -70,17 +85,36 @@ namespace EntityGraphQL.Schema
             object? argInstance = null;
             var validationErrors = new List<string>();
 
-            if (Arguments.Count > 0)
-            {
-                argInstance = ArgumentUtil.BuildArgumentsObject(Schema, Name, this, gqlRequestArgs ?? new Dictionary<string, object>(), Arguments.Values, ArgumentsType, variableParameter, docVariables, validationErrors);
-            }
-
             // add parameters and any DI services
             foreach (var p in method.GetParameters())
             {
                 if (p.GetCustomAttribute(typeof(MutationArgumentsAttribute)) != null || p.ParameterType.GetTypeInfo().GetCustomAttribute(typeof(MutationArgumentsAttribute)) != null)
                 {
+                    argInstance = ArgumentUtil.BuildArgumentsObject(Schema, Name, this, gqlRequestArgs ?? new Dictionary<string, object>(), Arguments.Values, ArgumentsType, variableParameter, docVariables, validationErrors);
                     allArgs.Add(argInstance!);
+                }
+                else if (docVariables != null && gqlRequestArgs != null && gqlRequestArgs.ContainsKey(p.Name))
+                {                    
+                    var argField = Arguments[p.Name];
+                    var value = ArgumentUtil.BuildArgumentFromMember(Schema, gqlRequestArgs ?? new Dictionary<string, object>(), argField.Name, argField.RawType, argField.DefaultValue, validationErrors);
+                    // this could be int to RequiredField<int>
+                    if (value != null && value.GetType() != argField.RawType)
+                    {
+                        value = ExpressionUtil.ChangeType(value, argField.RawType, Schema);
+                    }
+
+                    if (value == null)
+                    {
+                        var field = docVariables.GetType().GetField(p.Name);
+                        if (field != null)
+                        {
+                            value = field.GetValue(docVariables);
+                        }
+                    }
+
+                    argField.Validate(value, p.Name, validationErrors);
+
+                    allArgs.Add(value!);
                 }
                 else if (p.ParameterType == context.GetType())
                 {