+++ src/tests/EntityGraphQL.Tests/TestAbstractDataContext.cs
@@ -10,6 +10,14 @@ namespace EntityGraphQL.Tests
     public class TestAbstractDataContext
     {
         public List<Animal> Animals { get; set; } = new List<Animal>();
+        public List<Cat> Cats { get; set; } = new List<Cat>();
+        public List<Dog> Dogs { get; set; } = new List<Dog>();
+    }
+
+    public class TestAbstractDataContextNoAnimals
+    {
+        public List<Cat> Cats { get; set; } = new List<Cat>();
+        public List<Dog> Dogs { get; set; } = new List<Dog>();
     }
 
     public abstract class Animal